requirements.txt
Adafruit-SSD1306
AWSIoTPythonSDK
requests